Add the Local Request Options to the method signature
In the Axios targeted client, there is an additional parameter on all of the api action methods to pass in additional options for the config.

For my api I need to pass in per request nonce headers.  I added the localVarRequestOptions to the method signatures so that these could be passed in like they are in the react front end for the same api.